requirements.txt
python-dotenv==1.0.1
supabase==2.29.0
requests==2.32.3
openai==1.55.0
httpx==0.27.0
pydantic==2.10.5
websockets==15.0.1
